Upgrade rclone to current version

+++ tests/systemd/scripts/generate-retroarch-units.sh
@@ -109,12 +109,11 @@ function newPathUnitsFromDir() {
 
 [ $? = 0 ] || exit $?
 
-echo "Test 2: SUCCESS"
+echo "TEST 2 passed."
 
 ########
 # TEST 3
 ########
-##################################################################
 cat <<EOF >"${ARKLONE[retroarchCfg]}"
 savefile_directory = "/foo/bar"
 savefiles_in_content_dir = "false"
@@ -151,14 +150,20 @@ function newPathUnitsFromDir() {
 
 [ $? = 0 ] || exit $?
 
-echo "Test 3: SUCCESS"
+echo "TEST 3 passed."
 
 ########
 # TEST 4
 ########
-ARKLONE[retroarchContentRoot]="/foo/baz"
+# Skip test if run on ArkOS
+if [ "$(lsblk -f | grep "EASYROMS" | cut -d ' ' -f 2)" = "exfat" ]; then
+  echo "TEST 4 skipped."
 
-cat <<EOF >"${ARKLONE[retroarchCfg]}"
+# Run script
+else
+	ARKLONE[retroarchContentRoot]="/foo/baz"
+
+	cat <<EOF >"${ARKLONE[retroarchCfg]}"
 savefile_directory = "/foo/bar"
 savefiles_in_content_dir = "true"
 sort_savefiles_by_content_enable = "false"
@@ -169,33 +174,29 @@ sort_savestates_by_content_enable = "false"
 sort_savestates_enable = "false"
 EOF
 
-function newPathUnitsFromDir() {
-	# Function was called with correct local directory
-	[ "${1}" = "/foo/baz" ] || exit 64
+	function newPathUnitsFromDir() {
+		# Function was called with correct local directory
+		[ "${1}" = "/foo/baz" ] || exit 64
 
-	# Function was called with correct remote directory
-	[ "${2}" = "retroarch32" ] || exit 64
+		# Function was called with correct remote directory
+		[ "${2}" = "retroarch32/baz" ] || exit 64
 
-	# Function was called with correct subdir depth
-	[ "${3}" = "1" ] || exit 64
+		# Function was called with correct subdir depth
+		[ "${3}" = "1" ] || exit 64
 
-	# Function was called with makeRootUnit = true
-	[ "${4}" = "true" ] || exit 64
+		# Function was called with makeRootUnit = true
+		[ "${4}" = "true" ] || exit 64
 
-	# Function was called with correct filters
-	[ "${5}" = "retroarch-savefile|retroarch-savestate" ] || exit 64
+		# Function was called with correct filters
+		[ "${5}" = "retroarch-savefile|retroarch-savestate" ] || exit 64
 
-	# Function was called with correct ignore file
-	# @todo ArkOS-specific
-	[ "${6}" = "arkos-retroarch-content-root.ignore" ] || exit 64
-}
+		# Function was called with correct ignore file
+		# @todo ArkOS-specific
+		[ "${6}" = "${ARKLONE[ignoreDir]}/arkos-retroarch-content-root.ignore" ] || exit 64
+	}
 
-# Run script
-. "${ARKLONE[installDir]}/systemd/scripts/generate-retroarch-units.sh"
-
-[ $? = 0 ] || exit $?
-
-echo "Test 4: SUCCESS"
+		echo "TEST 4 passed."
+fi
 
 ########
 # TEST 5
@@ -217,7 +218,7 @@ function newPathUnitsFromDir() {
 # Run script
 . "${ARKLONE[installDir]}/systemd/scripts/generate-retroarch-units.sh" true
 
-echo "Test 5: SUCCESS"
+echo "TEST 5 passed."
 
 ##########
 # TEARDOWN
